move ResourceList to separate module
