+++ app/containers/ProjectPage.js
@@ -3,7 +3,8 @@ import marked from 'marked';
 import React from 'react';
 import { Link } from 'react-router';
 import request from 'superagent';
-import ResourceContainer from './ResourceContainer.js';
+import ResourceContainer from './ResourceContainer';
+import ResourceList from '../components/ResourceList';
 
 
 // Container for resources associated with the selected project
@@ -99,39 +100,5 @@ const ResourceForm = React.createClass({
   }
 });
 
-const ResourceList = React.createClass({
-
-  // render returns an array of Resource components by mapping the resource objects
-  // stored in this.props.data
-  render() {
-    const { data, getResources } = this.props;
-
-    const resourceNodes = data.map((resource) => {
-      const deleteResource = () => {
-        request
-          .delete(`/api/resources/${resource._id}`)
-          .end((err, res) => {
-            if (err || !res.ok) {
-              console.log(err);
-            } else {
-              getResources();
-            }
-          });
-      };
-
-      return (
-        <Resource key={resource._id} resource={resource} deleteResource={deleteResource}>
-          {resource.url}
-        </Resource>
-      );
-    });
-
-    return (
-      <div className="resourceList">
-        {resourceNodes}
-      </div>
-    );
-  }
-});
 
 export default ProjectPage;
